+++ Changes.md
@@ -1,3 +1,9 @@
+0.03
+====
+- Renamed players_list() to players()
+  - Old players() removed as it had no functionality
+- Logs returns array ref in scalar context now
+
 0.02
 ====
 - Lots of logic refractored to ask Player objects what the result is, such that different roles can change the result without the Games::Mafia object needing to be made aware
@@ -12,4 +18,4 @@
 
 0.01
 ====
-- Created *Wed Jun 27 12:03:48 2012*
+- Created *Wed, 27 Jun 2012 21:33:48 UTC*